divided the content into different components
