+++ code/src/Components/LikedThoughts.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const HeartIcon = (props) => (
+    <span >
+        {props.symbol}
+    </span>
+)
