+++ code/src/Components/ThoughtsForm.js
@@ -1,30 +1,11 @@
-import React, { useState } from 'react'
+import React from 'react'
 
 
-export const ThoughtsForm = () => {
-    const [newMessages, setNewMessages] = useState('')
-    const onNewMessagesChange = (event) => {
-        setNewMessages(event.target.value)
-    }
-    const onFormSubmit = (event) => {
-        event.preventDefault()
-         const options =  {
-                method: 'POST',
-                headers: {
-                    'content-Type': 'application/json'
-                },
-                body: JSON.stringify ({
-                    message : newMessages
-                })
-            }
-            fetch('https://happy-thoughts-technigo.herokuapp.com/thoughts', options)
-            .then(res => res.json())
-            .then(data => console.log(data))
-    }
+export const ThoughtsForm = ({newMessages, onNewMessagesChange, onFormSubmit }) => {
     return (
         <form onSubmit={onFormSubmit}>
             <h1>Welcome to write a message about you thoughts below</h1>
-            <textarea value={newMessages} onChange={e => onNewMessagesChange(e)}/>
+            <textarea value={newMessages} onChange={onNewMessagesChange}/>
             <button type='submit'>Submit form!</button>
         </form>
     )
