+++ code/src/Components/ThoughtsList.js
@@ -1,29 +1,17 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import moment from 'moment'
+import { HeartIcon } from 'Components/LikedThoughts'
 
-export const ThoughtsList = () => {
-  const [list, setList] = useState([])
-  const [loading,setLoading] = useState(false)
-
-    useEffect(() => {
-        fetchList()
-    }, [])
-     const fetchList = () => {
-        setLoading(true)
-        fetch('https://happy-thoughts-technigo.herokuapp.com/thoughts')
-        .then(res => res.json())
-        .then(data => setList(data))
-        .catch(error => console.error(error))
-        .finally(() => setLoading(false))
-        }
-        if (loading) {
-            return <h1>Loading in progress....</h1>
-        }
+export const ThoughtsList = ({list}) => {
     return (
         <section>
            {list.map(thoughts => (
                < div key={thoughts._id}>
                    <h4>{thoughts.message}</h4>
+                   <button>
+                       <HeartIcon symbol='❤️'/>
+                   </button>
+                   <p> x {thoughts.hearts}</p>
                    <p>{moment(thoughts.createdAt).fromNow()}</p>
                    </div>
            ))}
@@ -34,27 +22,4 @@ export const ThoughtsList = () => {
 
 
 
-{/*export const ThoughtsList = (messages) => {
-    const [message, setMessage] = useState ([])
 
-    useEffect (() => {
-        fetch('https://happy-thoughts-technigo.herokuapp.com/thoughts')
-        .then(res => {return res.json()})
-        .then(data => {
-            setMessage(data)
-    })
-}, []); 
-
-return (
-    <div>
-        {message.map(messages => {
-                <div key={messages._id}>
-                    <p>{messages.message}</p>
-                </div>
-        })}
-        {messages}
-
-    </div>
-)
-
-}*/}
